# public/openapi/agent-commerce-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Veyra Agent API",
    "version": "1.5.0",
    "description": "Machine API for verified Veyra workflows, Project 360 free discovery and explicit quoting, continuous trust monitoring, alerts, signed webhooks, and public trust badges."
  },
  "servers": [
    {
      "url": "https://agent-commerce-six.vercel.app",
      "description": "Arc Testnet Production Server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/agent/v1/reports/{reportId}": {
      "get": {
        "summary": "Retrieve structured report",
        "description": "Returns the completed structured report and Arc verification proofs. Supports `Accept: application/json` or `Accept: text/markdown`. Requires `results:read` scope.",
        "operationId": "getReport",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "description": "The report identifier (matches runId)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept",
            "in": "header",
            "required": false,
            "description": "Media type negotiation: `application/json` (default) or `text/markdown`",
            "schema": {
              "type": "string",
              "enum": ["application/json", "text/markdown"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Structured JSON report or formatted Markdown text",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/StructuredReportResponse" },
                    { "$ref": "#/components/schemas/AgentTrustReportResponse" },
                    { "$ref": "#/components/schemas/Project360ReportResponse" }
                  ]
                }
              },
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Report not ready yet (`report_not_ready`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "aac_live_...",
        "description": "Veyra Agent API Bearer credential token (`aac_live_...`)"
      }
    },
...
    "schemas": {
...
      "Project360ReportResponse": {
        "type": "object",
        "required": ["schema", "reportId", "workflow", "workflowType", "confirmedSources", "modules", "score", "coverage", "sections", "verification", "generatedAt"],
        "properties": {
          "schema": { "type": "string", "enum": ["veyra.project360.v1"] },
          "reportId": { "type": "string" },
          "workflow": { "type": "string", "enum": ["project_360"] },
          "workflowType": { "type": "string", "enum": ["project_360"] },
          "confirmedSources": { "type": "array", "items": { "type": "object" } },
          "modules": { "type": "array", "items": { "type": "object" } },
          "score": {
            "type": "object",
            "required": ["formulaVersion", "value", "confidencePercent", "confidence", "breakdown"],
            "properties": {
              "formulaVersion": { "type": "string", "enum": ["project360-score-v1"] },
              "value": { "type": "integer", "minimum": 0, "maximum": 100, "nullable": true },
              "confidencePercent": { "type": "integer", "minimum": 0, "maximum": 100 },
              "confidence": { "type": "string", "enum": ["high", "medium", "low", "insufficient"] },
              "breakdown": { "type": "array", "items": { "type": "object" } }
            }
          },
          "coverage": {
            "type": "object",
            "required": ["expected", "completed", "total", "status", "label"],
            "properties": {
              "expected": { "type": "integer", "minimum": 1, "maximum": 5 },
              "completed": { "type": "integer", "minimum": 0, "maximum": 5 },
              "total": { "type": "integer", "enum": [5] },
              "status": { "type": "string", "enum": ["complete", "partial", "limited", "failed"] },
              "label": { "type": "string" }
            }
          },
          "sections": {
            "type": "array",
            "minItems": 15,
            "maxItems": 15,
            "items": { "type": "object" }
          },
          "verification": { "type": "object" },
          "generatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "StructuredReportResponse": {
        "type": "object",
        "required": [
          "reportId",
          "workflow",
          "repository",
          "status",
          "executiveSummary",
          "projectPurpose",
          "technology",
          "activity",
          "strengths",
          "risks",
          "questionsBeforeAdoption",
          "confidence",
          "verification",
          "generatedAt"
        ],
        "properties": {
          "reportId": { "type": "string", "example": "job_01h9a8b7c6d5" },
          "workflow": { "type": "string", "example": "github_due_diligence" },
          "repository": {
            "type": "object",
            "nullable": true,
            "properties": {
              "fullName": { "type": "string", "example": "circlefin/agent-commerce" },
              "canonicalUrl": { "type": "string", "example": "https://github.com/circlefin/agent-commerce" }
            }
          },
          "status": { "type": "string", "example": "completed" },
          "verdict": {
            "type": "object",
            "nullable": true,
            "required": ["code", "label", "confidence", "summary", "reasons", "blockingFindings", "evidenceCoverage"],
            "properties": {
              "code": {
                "type": "string",
                "enum": ["proceed_with_standard_review", "proceed_with_conditions", "pause_for_review", "insufficient_evidence"]
              },
              "label": { "type": "string", "example": "Proceed with standard review" },
              "confidence": { "type": "string", "enum": ["high", "medium", "low"] },
              "summary": { "type": "string" },
              "reasons": { "type": "array", "items": { "type": "string" } },
              "blockingFindings": { "type": "array", "items": { "type": "string" } },
              "evidenceCoverage": {
                "type": "object",
                "properties": {
                  "assessedCategories": { "type": "number" },
                  "highConfidenceCategories": { "type": "number" },
                  "totalCategories": { "type": "number" },
                  "ratio": { "type": "number" }
                }
              }
            }
          },
          "executiveSummary": { "type": "string", "example": "Repository due diligence completed with strong governance and regular commit activity." },
          "projectPurpose": {
            "type": "object",
            "properties": {
              "summary": { "type": "string" },
              "primaryInterface": { "type": "string" },
              "developmentStage": { "type": "string" },
              "targetUsers": { "type": "string" },
              "capabilities": { "type": "array", "items": { "type": "string" } }
            }
          },
          "technology": {
            "type": "object",
            "properties": {
              "primaryLanguage": { "type": "string", "example": "TypeScript" },
              "frameworks": { "type": "array", "items": { "type": "string" }, "example": ["Next.js", "React"] },
              "hasWorkflows": { "type": "boolean", "example": true },
              "workflowCount": { "type": "number", "example": 3 }
            }
          },
          "activity": {
            "type": "object",
            "properties": {
              "commitCount30d": { "type": "number", "example": 42 },
              "commitCount90d": { "type": "number", "example": 120 },
              "commitCount180d": { "type": "number", "example": 250 },
              "lastCommitAt": { "type": "string", "example": "2026-07-24T18:00:00Z" }
            }
          },
          "strengths": { "type": "array", "items": { "type": "string" }, "example": ["Comprehensive test suite", "Active maintainers"] },
          "risks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": { "type": "string", "example": "low_test_coverage" },
                "title": { "type": "string", "example": "Moderate Test Coverage" },
                "severity": { "type": "string", "example": "low" },
                "description": { "type": "string", "example": "Some helper functions lack unit tests." },
                "impact": { "type": "string", "example": "Potential edge case bugs in error handlers." }
              }
            }
          },
          "questionsBeforeAdoption": { "type": "array", "items": { "type": "string" }, "example": ["What is the long-term maintenance roadmap?"] },
          "confidence": { "type": "string", "example": "high" },
          "verification": {
            "type": "object",
            "properties": {
              "status": { "type": "string", "example": "verified" },
              "network": { "type": "string", "example": "arc-testnet" },
              "proofs": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "receiptId": { "type": "string", "example": "rcpt_01h9a8b7" },
                    "txHash": { "type": "string", "nullable": true, "example": "0xabcd1234ef567890" },
                    "status": { "type": "string", "example": "verified" },
                    "explorerUrl": { "type": "string", "example": "https://testnet.arcscan.app/tx/0xabcd1234ef567890" }
                  }
                }
              }
            }
          },
          "generatedAt": { "type": "string", "format": "date-time", "example": "2026-07-24T21:05:00.000Z" }
        }
      },
      "AgentTrustReportResponse": {
        "type": "object",
        "required": [
          "kind",
          "version",
          "workflowType",
          "reportId",
          "input",
          "subject",
          "trustScore",
          "executiveSummary",
          "identity",
          "codeIntelligence",
          "executionReliability",
          "paymentsAndReceipts",
          "services",
          "contractTransparency",
          "endpointAvailability",
          "evidenceBackedStrengths",
          "risksAndReviewItems",
          "questionsBeforeIntegration",
          "dataFreshness",
          "unavailableSources",
          "limitations",
          "verification",
          "generatedAt"
        ],
        "properties": {
          "kind": { "type": "string", "enum": ["agent_trust_report"] },
          "version": { "type": "integer", "enum": [1] },
          "workflowType": { "type": "string", "enum": ["agent_trust_report"] },
          "reportId": { "type": "string" },
          "input": { "$ref": "#/components/schemas/AgentTrustReportInput" },
          "subject": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "agentId": { "type": "string", "nullable": true },
              "wallet": { "type": "string", "nullable": true },
              "repository": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "fullName": { "type": "string" },
                  "canonicalUrl": { "type": "string" }
                }
              }
            }
          },
          "trustScore": {
            "type": "object",
            "required": ["overall", "status", "categories", "excludedCategories"],
            "properties": {
              "overall": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100,
                "nullable": true,
                "description": "Deterministic confidence-weighted score; null when fewer than two categories are scorable"
              },
              "status": {
                "type": "string",
                "enum": ["strong_signals", "review_recommended", "high_attention", "limited_data"]
              },
              "categories": {
                "type": "object",
                "additionalProperties": {
                  "type": "object",
                  "properties": {
                    "score": { "type": "integer", "minimum": 0, "maximum": 100, "nullable": true },
                    "confidence": { "type": "string", "enum": ["high", "medium", "low"] },
                    "evidenceCount": { "type": "integer" },
                    "summary": { "type": "string" },
                    "positiveSignals": { "type": "array", "items": { "type": "object" } },
                    "reviewItems": { "type": "array", "items": { "type": "object" } }
                  }
                }
              },
              "excludedCategories": { "type": "array", "items": { "type": "string" } }
            }
          },
          "executiveSummary": { "type": "array", "items": { "type": "string" } },
          "identity": { "type": "object" },
          "codeIntelligence": { "type": "object" },
          "executionReliability": { "type": "object" },
          "paymentsAndReceipts": { "type": "object" },
          "services": { "type": "object" },
          "contractTransparency": { "type": "object" },
          "endpointAvailability": { "type": "object" },
          "evidenceBackedStrengths": { "type": "array", "items": { "type": "object" } },
          "risksAndReviewItems": { "type": "array", "items": { "type": "object" } },
          "questionsBeforeIntegration": { "type": "array", "items": { "type": "string" } },
          "dataFreshness": { "type": "array", "items": { "type": "object" } },
          "unavailableSources": { "type": "array", "items": { "type": "string" } },
          "limitations": { "type": "array", "items": { "type": "string" } },
          "verification": {
            "type": "object",
            "required": ["status", "verifiedOnArc", "network", "chainId", "reportHash", "proofs"],
            "properties": {
              "status": { "type": "string", "enum": ["verified", "verification_pending", "verification_failed"] },
              "verifiedOnArc": {
                "type": "boolean",
                "description": "True only when a verified Arc proof response hash exactly matches reportHash"
              },
              "network": { "type": "string", "enum": ["arc-testnet"] },
              "chainId": { "type": "integer", "enum": [5042002] },
              "reportHash": {
                "type": "string",
                "pattern": "^0x[0-9a-f]{64}$",
                "description": "Deterministic SHA-256 hash of the canonical report evidence and score snapshot"
              },
              "proofs": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "receiptId": { "type": "string" },
                    "status": { "type": "string", "enum": ["pending", "verified", "failed"] },
                    "transactionHash": { "type": "string", "nullable": true },
                    "explorerUrl": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          },
          "generatedAt": { "type": "string", "format": "date-time" }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid Bearer token credential",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Credential scope denied or workflow disabled",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      },
...
    }
  }
}